minor code formatting fixes
