+++ GiftTasteHelper/Framework/Controllers/GiftDataProvider.cs
@@ -1,9 +1,7 @@
-﻿using StardewValley;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using StardewValley;
 
 namespace GiftTasteHelper.Framework
 {
