+++ GiftTasteHelper/Framework/Controllers/GiftDatabase.cs
@@ -1,10 +1,6 @@
-﻿using StardewModdingAPI;
-using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Newtonsoft.Json;
+using StardewModdingAPI;
 
 namespace GiftTasteHelper.Framework
 {
@@ -101,7 +97,7 @@ namespace GiftTasteHelper.Framework
             }
             return new int[] { };
         }
-    }   
+    }
 
     /// <summary>A gift database that is stored on disk.</summary>
     internal class StoredGiftDatabase : GiftDatabase
