+++ GiftTasteHelper/Framework/Controllers/GiftDrawDataProvider.cs
@@ -1,8 +1,8 @@
-﻿using Microsoft.Xna.Framework;
-using StardewValley;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Microsoft.Xna.Framework;
+using StardewValley;
 
 namespace GiftTasteHelper.Framework
 {
