+++ GiftTasteHelper/GiftTasteHelper.csproj
@@ -48,6 +48,7 @@
     <Compile Include="Framework\Controllers\IGiftDatabase.cs" />
     <Compile Include="Framework\Controllers\IGiftDrawDataProvider.cs" />
     <Compile Include="Framework\Controllers\IGiftMonitor.cs" />
+    <Compile Include="Framework\Models\GiftConfig.cs" />
     <Compile Include="Framework\Models\GiftDatabaseModel.cs" />
     <Compile Include="Framework\GiftHelperTypes.cs" />
     <Compile Include="GiftTasteHelper.cs" />
