+++ GiftTasteHelper/i18n/default.json
@@ -1,4 +1,4 @@
 ﻿{
   "tooltip.title.favorite": "Favorite Gifts",
-  "tooltip.title.none":  "No Known Gifts"
+  "tooltip.title.none": "No Known Gifts"
 }
