+++ GiftTasteHelper/i18n/pt.json
